+++ packages/components/src/RichTooltip/RichError/RichError.component.js
