+++ packages/components/src/RichTooltip/RichError/RichError.component.test.js
