+++ packages/components/src/RichTooltip/RichError/RichError.scss
