+++ packages/components/src/RichTooltip/RichError/index.js
