+++ packages/components/src/RichTooltip/RichLayout/RichLayout.component.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import Inject from '../Inject';
+import Inject from '../../Inject';
 import theme from './RichLayout.scss';
 
 export default function RichLayout(props) {
