+++ packages/components/src/RichTooltip/RichLayout/RichLayout.scss
