+++ packages/components/src/RichTooltip/RichLayout/RichLayout.test.js
