+++ packages/components/src/index.js
@@ -53,8 +53,8 @@ import Typeahead from './Typeahead';
 import VirtualizedList from './VirtualizedList';
 import WithDrawer from './WithDrawer';
 import Inject from './Inject';
-import RichLayout from './RichLayout';
-import RichError from './RichLayout/RichError';
+import RichLayout from './RichTooltip/RichLayout';
+import RichError from './RichTooltip/RichError';
 
 const {
 	Alert,
